+++ docs/upgrading.md
@@ -256,45 +256,3 @@ To use default attribute persistence, you need to pass in a
 `PersistentStorageDelegate` to `CodegenDataModelProviderInstance`. See example
 changes in [36658](https://github.com/project-chip/connectedhomeip/pull/36658)
 ).
-
-### DiagnosticLogsCluster API Changes
-
-The Diagnostic Logs cluster is a device-wide singleton. There is no need to
-manage delegates on a per-endpoint basis. To reflect this, the API has been
-simplified.
-
-What Changed
-
-#### SetDelegate() API Updated
-
-Old API (deprecated):
-
-```
-chip::app::Clusters::DiagnosticLogs::DiagnosticLogsServer::SetDiagnosticLogsProviderDelegate(endpointId, delegate);
-```
-
--   Required passing an EndpointId.
--   Misleading, since Diagnostic Logs is not endpoint-scoped.
-
-New API:
-
-```
-chip::app::Clusters::DiagnosticLogsCluster::Instance().SetDelegate(delegate);
-```
-
--   No endpointId required.
--   Clearer: directly sets the global delegate on the singleton cluster.
-
-#### Backward Compatibility
-
-For migration ease:
-
-The old method
-
-```
-DiagnosticLogsServer::SetDiagnosticLogsProviderDelegate(endpointId, delegate)
-```
-
-still exists.
-
-callers should migrate to the new Instance().SetDelegate() API.
